Revise/unify buttons, cleanup default ui element code, add panel button
