+++ static/scripts/mvc/app/app-panel-history.js
@@ -1,2 +1,2 @@
-define(["utils/utils","mvc/history/options-menu","mvc/history/history-panel-edit-current"],function(a,b,c){return Backbone.View.extend({initialize:function(d){this.options=a.merge(d,{}),this.setElement(this._template()),this.$("#history-refresh-button").on("click",function(){top.Galaxy&&top.Galaxy.currHistoryPanel&&top.Galaxy.currHistoryPanel.loadCurrentHistory()});var e=b(this.$("#history-options-button"),{anonymous:d.user.valid&&"true"||"false",purgeAllowed:Galaxy.config.allow_user_dataset_purge&&"true"||"false",root:Galaxy.root});Galaxy.historyOptionsMenu=e;var f=new c.CurrentHistoryPanel({el:this.$("#current-history-panel"),purgeAllowed:Galaxy.config.allow_user_dataset_purge,linkTarget:"galaxy_main",$scrollContainer:function(){return this.$el.parent()}});f.connectToQuotaMeter(Galaxy.quotaMeter),f.listenToGalaxy(Galaxy),f.loadCurrentHistory(),Galaxy.currHistoryPanel=f;new Panel({panel:this.$el,center:this.$("#center"),drag:this.$(".unified-panel-footer > .drag"),toggle:this.$(".unified-panel-footer > .panel-collapse"),right:!0})},_template:function(){return'<div id="right"><div class="unified-panel-header" unselectable="on"><div class="unified-panel-header-inner history-panel-header"><div style="float: right"><a id="history-refresh-button" class="panel-header-button" href="javascript:void(0)" title="Refresh history"><span class="fa fa-refresh"/></a><a id="history-options-button" class="panel-header-button" href="'+Galaxy.root+'root/history_options" target="galaxy_main" title="History options"><span class="fa fa-cog"/></a><a id="history-view-multi-button" class="panel-header-button" href="'+Galaxy.root+'history/view_multiple" title="View all histories"><span class="fa fa-columns"/></a></div><div class="panel-header-text">History</div></div></div><div class="unified-panel-body"><div id="current-history-panel" class="history-panel"/></div><div class="unified-panel-footer"><div class="panel-collapse right"/><div class="drag"/></div></div>'}})});
+define(["utils/utils","mvc/ui/ui-misc","mvc/history/options-menu","mvc/history/history-panel-edit-current"],function(a,b,c,d){return Backbone.View.extend({initialize:function(e){this.options=a.merge(e,{}),this.setElement(this._template());var f=c(this.$("#history-options-button"),{anonymous:e.user.valid&&"true"||"false",purgeAllowed:Galaxy.config.allow_user_dataset_purge&&"true"||"false",root:Galaxy.root});Galaxy.historyOptionsMenu=f;var g=new d.CurrentHistoryPanel({el:this.$el,purgeAllowed:Galaxy.config.allow_user_dataset_purge,linkTarget:"galaxy_main",$scrollContainer:function(){return this.$el.parent()}});this.components={header:{title:"History",buttons:[new b.ButtonLink({id:"history-refresh-button",title:"Refresh history",cls:"panel-header-button",icon:"fa fa-refresh",onclick:function(){top.Galaxy&&top.Galaxy.currHistoryPanel&&top.Galaxy.currHistoryPanel.loadCurrentHistory()}})]}},g.connectToQuotaMeter(Galaxy.quotaMeter),g.listenToGalaxy(Galaxy),g.loadCurrentHistory(),Galaxy.currHistoryPanel=g},_template:function(){return'<div id="current-history-panel" class="history-panel"/>'}})});
 //# sourceMappingURL=../../../maps/mvc/app/app-panel-history.js.map
