+++ static/scripts/mvc/app/app-panel-tool.js
@@ -1,2 +1,2 @@
-define(["utils/utils","mvc/tools"],function(a,b){return Backbone.View.extend({initialize:function(c){if(this.options=a.merge(c,{}),this.setElement(this._template()),Galaxy.user.valid||!Galaxy.config.require_login){var d=new b.ToolSearch({spinner_url:c.spinner_url,search_url:c.search_url,hidden:!1}),e=new b.ToolCollection(c.toolbox),f=new b.ToolPanel({tool_search:d,tools:e,layout:c.toolbox_in_panel});tool_panel_view=new b.ToolPanelView({model:f}),Galaxy.toolPanel=f,f.get("layout").size()>0&&(tool_panel_view.render(),this.$(".toolMenu").show()),this.$(".toolMenuContainer").prepend(tool_panel_view.$el),this.$("a[minsizehint]").click(function(){parent.handle_minwidth_hint&&parent.handle_minwidth_hint($(this).attr("minsizehint"))})}new Panel({panel:this.$el,center:this.$("#center"),drag:this.$(".unified-panel-footer > .drag"),toggle:this.$(".unified-panel-footer > .panel-collapse")})},_template:function(){return'<div id="left"><div class="unified-panel-header" unselectable="on"><div class="unified-panel-header-inner with-upload-button"> Tools </div></div><div class="unified-panel-body"><div class="toolMenuContainer"><div class="toolMenu" style="display: none"><div id="search-no-results" style="display: none; padding-top: 5px"><em><strong>Search did not match any tools.</strong></em></div></div></div></div><div class="unified-panel-footer"><div class="panel-collapse left"/><div class="drag"/></div></div>'}})}),_;
+define(["utils/utils","mvc/tools"],function(a,b){return Backbone.View.extend({initialize:function(c){if(this.options=a.merge(c,{}),this.setElement(this._template()),Galaxy.user.valid||!Galaxy.config.require_login){var d=new b.ToolSearch({spinner_url:c.spinner_url,search_url:c.search_url,hidden:!1}),e=new b.ToolCollection(c.toolbox),f=new b.ToolPanel({tool_search:d,tools:e,layout:c.toolbox_in_panel});tool_panel_view=new b.ToolPanelView({model:f}),Galaxy.toolPanel=f,f.get("layout").size()>0&&(tool_panel_view.render(),this.$(".toolMenu").show()),this.$(".toolMenuContainer").prepend(tool_panel_view.$el),this.$("a[minsizehint]").click(function(){parent.handle_minwidth_hint&&parent.handle_minwidth_hint($(this).attr("minsizehint"))})}},_template:function(){return'<div class="unified-panel"><div class="unified-panel-header" unselectable="on"><div class="unified-panel-header-inner with-upload-button"> Tools </div></div><div class="unified-panel-body"><div class="toolMenuContainer"><div class="toolMenu" style="display: none"><div id="search-no-results" style="display: none; padding-top: 5px"><em><strong>Search did not match any tools.</strong></em></div></div></div></div><div class="unified-panel-footer">testsadasdsa</div></div>'}})});
 //# sourceMappingURL=../../../maps/mvc/app/app-panel-tool.js.map
