+++ static/scripts/mvc/app/app-view.js
@@ -1,2 +1,2 @@
-define(["utils/utils","mvc/ui/ui-portlet","mvc/ui/ui-misc","galaxy.masthead","galaxy.menu","mvc/ui/ui-modal","galaxy.frame","mvc/upload/upload-view","mvc/user/user-model","mvc/user/user-quotameter","mvc/app/app-panel-history","mvc/app/app-panel-tool","mvc/app/app-panel-center"],function(a,b,c,d,e,f,g,h,i,j,k,l,m){return Backbone.View.extend({initialize:function(b){this.options=a.merge(b,{}),this.setElement(this._template(b)),ensure_dd_helper(),$("body").append(this.$el),Galaxy.currUser||(Galaxy.currUser=new i.User(b.user.json)),Galaxy.masthead||(Galaxy.masthead=new d.GalaxyMasthead(b),Galaxy.modal=new f.View,Galaxy.frame=new g.GalaxyFrame,Galaxy.menu=new e.GalaxyMenu({masthead:Galaxy.masthead,config:b}),Galaxy.quotaMeter=new j.UserQuotaMeter({model:Galaxy.currUser,el:Galaxy.masthead.$(".quota-meter-container")}).render());var c=new k(b),n=new l(b),o=new m;this.$("#right-panel").append(c.$el),this.$("#left-panel").append(n.$el),this.$("#center-panel").append(o.$el),Galaxy.upload=new h(b)},_template:function(){return'<div id="everything" style="position: absolute; top: 0; left: 0; width: 100%; height: 100%;"><div id="masthead" class="navbar navbar-fixed-top navbar-inverse"/><div id="left-panel"/><div id="center-panel" class="inbound"/><div id="right-panel"/></div>'}})});
+define(["utils/utils","mvc/ui/ui-portlet","mvc/ui/ui-misc","galaxy.masthead","galaxy.menu","mvc/ui/ui-modal","galaxy.frame","mvc/upload/upload-view","mvc/user/user-model","mvc/user/user-quotameter","mvc/app/app-panel-history","mvc/app/app-panel-tool","mvc/app/app-panel-center"],function(a,b,c,d,e,f,g,h,i,j,k,l,m){return Backbone.View.extend({initialize:function(b){this.options=a.merge(b,{}),this.setElement(this._template(b)),ensure_dd_helper(),$("body").append(this.$el),Galaxy.currUser||(Galaxy.currUser=new i.User(b.user.json)),Galaxy.masthead||(Galaxy.masthead=new d.GalaxyMasthead(b),Galaxy.modal=new f.View,Galaxy.frame=new g.GalaxyFrame,Galaxy.menu=new e.GalaxyMenu({masthead:Galaxy.masthead,config:b}),Galaxy.quotaMeter=new j.UserQuotaMeter({model:Galaxy.currUser,el:Galaxy.masthead.$(".quota-meter-container")}).render());var c=new l(b),n=new k(b),o=new m;this._buildPanel("#right",n),this.$("#left").append(c.$el.children()),this.$("#center").append(o.$el),Galaxy.upload=new h(b);new Panel({center:this.$("#center"),panel:this.$("#left"),drag:this.$("#left .unified-panel-footer > .drag"),toggle:this.$("#left .unified-panel-footer > .panel-collapse")})},_buildPanel:function(a,b){var c=this.$(a),d=b.components;c.find(".panel-header-text").html(d.header.title);for(var e in d.header.buttons)c.find(".panel-header-buttons").append(d.header.buttons[e].$el);c.find(".unified-panel-body").append(b.$el),new Panel({center:this.$("#center"),panel:this.$(a),drag:this.$(a+".unified-panel-footer > .drag"),toggle:this.$(a+".unified-panel-footer > .panel-collapse"),right:!0})},_template:function(){return'<div id="everything" style="position: absolute; top: 0; left: 0; width: 100%; height: 100%;"><div id="masthead" class="navbar navbar-fixed-top navbar-inverse"/><div id="left"/><div id="center" class="inbound"/><div id="right"><div class="unified-panel-header" unselectable="on"><div class="unified-panel-header-inner history-panel-header"><div class="panel-header-buttons" style="float: right"/><div class="panel-header-text"/></div></div><div class="unified-panel-body"/><div class="unified-panel-footer"><div class="panel-collapse right"/><div class="drag"/></div></div></div>'}})});
 //# sourceMappingURL=../../../maps/mvc/app/app-view.js.map
