--- static/scripts/mvc/ui/ui-button-check.js
@@ -1,2 +1,0 @@
-define(["utils/utils"],function(a){return Backbone.View.extend({optionsDefault:{title:"Select/Unselect all",icons:["fa fa-square-o","fa fa-minus-square-o","fa fa-check-square-o"],value:0},initialize:function(b){this.options=a.merge(b,this.optionsDefault),this.setElement(this._template()),this.$title=this.$(".title"),this.$icon=this.$(".icon"),this.value(this.options.value),this.$title.html(this.options.title);var c=this;this.$el.on("click",function(){c.current=0===c.current&&2||0,c.value(c.current),c.options.onclick&&c.options.onclick()})},value:function(a,b){return void 0!==a&&(b&&0!==a&&(a=a!==b&&1||2),this.current=a,this.$icon.removeClass().addClass("icon").addClass(this.options.icons[a]),this.options.onchange&&this.options.onchange(a)),this.current},_template:function(){return'<div class="ui-button-check" ><span class="icon"/><span class="title"/></div>'}})});
-//# sourceMappingURL=../../../maps/mvc/ui/ui-button-check.js.map
