--- static/scripts/mvc/ui/ui-button-menu.js
@@ -1,2 +1,0 @@
-define(["utils/utils"],function(a){return Backbone.View.extend({optionsDefault:{id:"",title:"",floating:"right",pull:"right",icon:null,onclick:null,cls:"ui-button-icon ui-button-menu",tooltip:"",target:"",href:"",onunload:null,visible:!0,tag:""},$menu:null,initialize:function(b){this.options=a.merge(b,this.optionsDefault),this.setElement($(this._template(this.options)));var c=$(this.el).find(".root"),d=this;c.on("click",function(a){$(".tooltip").hide(),a.preventDefault(),d.options.onclick&&d.options.onclick()}),this.options.visible||this.hide(),c.tooltip({title:b.tooltip,placement:"bottom"})},show:function(){$(this.el).show()},hide:function(){$(this.el).hide()},addMenu:function(b){var c={title:"",target:"",href:"",onclick:null,divider:!1,icon:null,cls:"button-menu btn-group"};c=a.merge(b,c),this.$menu||($(this.el).append(this._templateMenu()),this.$menu=$(this.el).find(".menu"));var d=$(this._templateMenuItem(c));d.on("click",function(a){a.preventDefault(),c.onclick&&c.onclick()}),this.$menu.append(d),c.divider&&this.$menu.append($(this._templateDivider()))},_templateMenuItem:function(a){var b='<li><a class="dropdown-item" href="'+a.href+'" target="'+a.target+'">';return a.icon&&(b+='<i class="fa '+a.icon+'"></i>'),b+=" "+a.title+"</a></li>"},_templateMenu:function(){return'<ul class="menu dropdown-menu pull-'+this.options.pull+'" role="menu"></ul>'},_templateDivider:function(){return'<li class="divider"></li>'},_template:function(a){var b="",c="";a.title?b="width: auto;":c="margin: 0px;";var d='<div id="'+a.id+'" style="float: '+a.floating+"; "+b+'" class="dropdown '+a.cls+'"><div class="root button dropdown-toggle" data-toggle="dropdown" style="'+c+'"><i class="icon fa '+a.icon+'"/>';return a.title&&(d+='&nbsp;<span class="title">'+a.title+"</span>"),d+="</div></div>"}})});
-//# sourceMappingURL=../../../maps/mvc/ui/ui-button-menu.js.map
