+++ static/scripts/mvc/ui/ui-misc.js
@@ -1,2 +1,2 @@
-define(["utils/utils","mvc/ui/ui-select-default","mvc/ui/ui-slider","mvc/ui/ui-options","mvc/ui/ui-drilldown","mvc/ui/ui-button-menu","mvc/ui/ui-button-check","mvc/ui/ui-modal"],function(a,b,c,d,e,f,g,h){var i=Backbone.View.extend({optionsDefault:{url:"",cls:""},initialize:function(b){this.options=a.merge(b,this.optionsDefault),this.setElement(this._template(this.options))},_template:function(a){return'<img class="ui-image '+a.cls+'" src="'+a.url+'"/>'}}),j=Backbone.View.extend({optionsDefault:{title:"",cls:""},initialize:function(b){this.options=a.merge(b,this.optionsDefault),this.setElement(this._template(this.options))},title:function(a){this.$el.html(a)},_template:function(a){return'<label class="ui-label '+a.cls+'">'+a.title+"</label>"},value:function(){return options.title}}),k=Backbone.View.extend({optionsDefault:{floating:"right",icon:"",tooltip:"",placement:"bottom",title:"",cls:""},initialize:function(b){this.options=a.merge(b,this.optionsDefault),this.setElement(this._template(this.options)),$(this.el).tooltip({title:b.tooltip,placement:"bottom"})},_template:function(a){return'<div><span class="fa '+a.icon+'" class="ui-icon"/>&nbsp;'+a.title+"</div>"}}),l=Backbone.View.extend({optionsDefault:{id:null,title:"",floating:"right",cls:"ui-button btn btn-default",icon:""},initialize:function(b){this.options=a.merge(b,this.optionsDefault),this.options.id=this.options.id||a.uid(),this.setElement(this._template(this.options)),$(this.el).on("click",function(){$(".tooltip").hide(),b.onclick&&b.onclick()}),$(this.el).tooltip({title:b.tooltip,placement:"bottom"})},disable:function(){this.$el.addClass("disabled"),this.disabled=!0},enable:function(){this.$el.removeClass("disabled"),this.disabled=!1},wait:function(){this.$el.removeClass(this.options.cls).addClass("btn btn-info").prop("disabled",!0),this.$(".icon").removeClass(this.options.icon).addClass("fa-spinner fa-spin"),this.$(".title").html("Sending...")},unwait:function(){this.$el.removeClass("btn btn-info").addClass(this.options.cls).prop("disabled",!1),this.$(".icon").removeClass("fa-spinner fa-spin").addClass(this.options.icon),this.$(".title").html(this.options.title)},_template:function(a){var b='<button id="'+a.id+'" type="submit" style="float: '+a.floating+';" type="button" class="'+a.cls+'">';return a.icon&&(b+='<i class="icon fa '+a.icon+'"></i>&nbsp;'),b+='<span class="title">'+a.title+"</span></button>"}}),m=Backbone.View.extend({optionsDefault:{id:a.uid(),title:"",floating:"right",cls:"ui-button-icon",icon:"",tooltip:"",onclick:null},initialize:function(b){this.options=a.merge(b,this.optionsDefault),this.setElement(this._template(this.options)),this.$button=this.$el.find(".button");var c=this;$(this.el).on("click",function(){$(".tooltip").hide(),b.onclick&&!c.disabled&&b.onclick()}),this.$button.tooltip({title:b.tooltip,placement:"bottom"})},disable:function(){this.$button.addClass("disabled"),this.disabled=!0},enable:function(){this.$button.removeClass("disabled"),this.disabled=!1},setIcon:function(a){this.$("i").removeClass(this.options.icon).addClass(a),this.options.icon=a},_template:function(a){var b="";a.title&&(b="width: auto;");var c='<div id="'+a.id+'" style="float: '+a.floating+"; "+b+'" class="'+a.cls+'"><div class="button">';return c+=a.title?'<i class="icon fa '+a.icon+'"/>&nbsp;<span class="title">'+a.title+"</span>":'<i class="icon fa '+a.icon+'"/>',c+="</div></div>"}}),n=Backbone.View.extend({optionsDefault:{title:"",cls:""},initialize:function(b){this.options=a.merge(b,this.optionsDefault),this.setElement(this._template(this.options)),$(this.el).on("click",b.onclick)},_template:function(a){return'<div><a href="javascript:void(0)" class="ui-anchor '+a.cls+'">'+a.title+"</a></div>"}}),o=Backbone.View.extend({optionsDefault:{message:null,status:"info",cls:"",persistent:!1},initialize:function(b){this.options=a.merge(b,this.optionsDefault),this.setElement('<div class="'+this.options.cls+'"/>'),this.options.message&&this.update(this.options)},update:function(b){if(this.options=a.merge(b,this.optionsDefault),""!=b.message){if(this.$el.html(this._template(this.options)),this.$el.find(".alert").append(b.message),this.$el.fadeIn(),this.timeout&&window.clearTimeout(this.timeout),!b.persistent){var c=this;this.timeout=window.setTimeout(function(){c.$el.is(":visible")?c.$el.fadeOut():c.$el.hide()},3e3)}}else this.$el.fadeOut()},_template:function(a){return'<div class="ui-message alert alert-'+a.status+'"/>'}}),p=Backbone.View.extend({optionsDefault:{onclick:null,searchword:""},initialize:function(b){this.options=a.merge(b,this.optionsDefault),this.setElement(this._template(this.options));var c=this;this.options.onclick&&this.$el.on("submit",function(){var a=c.$el.find("#search");c.options.onclick(a.val())})},_template:function(a){return'<div class="ui-search"><form onsubmit="return false;"><input id="search" class="form-control input-sm" type="text" name="search" placeholder="Search..." value="'+a.searchword+'"><button type="submit" class="btn search-btn"><i class="fa fa-search"></i></button></form></div>'}}),q=Backbone.View.extend({optionsDefault:{type:"text",placeholder:"",disabled:!1,visible:!0,cls:"",area:!1},initialize:function(b){this.options=a.merge(b,this.optionsDefault),this.setElement(this._template(this.options)),void 0!==this.options.value&&this.value(this.options.value),this.options.disabled&&this.$el.prop("disabled",!0),this.options.visible||this.$el.hide();var c=this;this.$el.on("input",function(){c.options.onchange&&c.options.onchange(c.$el.val())})},value:function(a){return void 0!==a&&this.$el.val(a),this.$el.val()},_template:function(a){return a.area?'<textarea id="'+a.id+'" class="ui-textarea '+a.cls+'"></textarea>':'<input id="'+a.id+'" type="'+a.type+'" value="'+a.value+'" placeholder="'+a.placeholder+'" class="ui-input '+a.cls+'">'}}),r=Backbone.View.extend({initialize:function(a){this.options=a,this.setElement(this._template(this.options)),void 0!==this.options.value&&this.value(this.options.value)},value:function(a){return void 0!==a&&this.$("hidden").val(a),this.$("hidden").val()},_template:function(a){var b='<div id="'+a.id+'" >';return a.info&&(b+="<div>"+a.info+"</div>"),b+='<hidden value="'+a.value+'"/></div>'}});return{Anchor:n,Button:l,ButtonIcon:m,ButtonCheck:g,ButtonMenu:f,Icon:k,Image:i,Input:q,Label:j,Message:o,Modal:h,RadioButton:d.RadioButton,Checkbox:d.Checkbox,Radio:d.Radio,Searchbox:p,Select:b,Hidden:r,Slider:c,Drilldown:e}});
+define(["utils/utils","mvc/ui/ui-select-default","mvc/ui/ui-slider","mvc/ui/ui-options","mvc/ui/ui-drilldown","mvc/ui/ui-buttons","mvc/ui/ui-modal"],function(a,b,c,d,e,f,g){var h=Backbone.View.extend({initialize:function(b){this.options=a.merge(b,{url:"",cls:""}),this.setElement(this._template(this.options))},_template:function(a){return'<img class="ui-image '+a.cls+'" src="'+a.url+'"/>'}}),i=Backbone.View.extend({initialize:function(b){this.options=a.merge(b,{title:"",cls:""}),this.setElement(this._template(this.options))},title:function(a){this.$el.html(a)},value:function(){return options.title},_template:function(a){return'<label class="ui-label '+a.cls+'">'+a.title+"</label>"}}),j=Backbone.View.extend({initialize:function(b){this.options=a.merge(b,{floating:"right",icon:"",tooltip:"",placement:"bottom",title:"",cls:""}),this.setElement(this._template(this.options)),$(this.el).tooltip({title:b.tooltip,placement:"bottom"})},_template:function(a){return'<div><span class="fa '+a.icon+'" class="ui-icon"/>&nbsp;'+a.title+"</div>"}}),k=Backbone.View.extend({initialize:function(b){this.options=a.merge(b,{title:"",cls:""}),this.setElement(this._template(this.options)),$(this.el).on("click",b.onclick)},_template:function(a){return'<div><a href="javascript:void(0)" class="ui-anchor '+a.cls+'">'+a.title+"</a></div>"}}),l=Backbone.View.extend({initialize:function(b){this.options=a.merge(b,{message:null,status:"info",cls:"",persistent:!1}),this.setElement('<div class="'+this.options.cls+'"/>'),this.options.message&&this.update(this.options)},update:function(b){if(this.options=a.merge(b,this.optionsDefault),""!=b.message){if(this.$el.html(this._template(this.options)),this.$el.find(".alert").append(b.message),this.$el.fadeIn(),this.timeout&&window.clearTimeout(this.timeout),!b.persistent){var c=this;this.timeout=window.setTimeout(function(){c.$el.is(":visible")?c.$el.fadeOut():c.$el.hide()},3e3)}}else this.$el.fadeOut()},_template:function(a){return'<div class="ui-message alert alert-'+a.status+'"/>'}}),m=Backbone.View.extend({initialize:function(b){this.options=a.merge(b,{onclick:null,searchword:""}),this.setElement(this._template(this.options));var c=this;this.options.onclick&&this.$el.on("submit",function(){var a=c.$el.find("#search");c.options.onclick(a.val())})},_template:function(a){return'<div class="ui-search"><form onsubmit="return false;"><input id="search" class="form-control input-sm" type="text" name="search" placeholder="Search..." value="'+a.searchword+'"><button type="submit" class="btn search-btn"><i class="fa fa-search"></i></button></form></div>'}}),n=Backbone.View.extend({initialize:function(b){this.options=a.merge(b,{type:"text",placeholder:"",disabled:!1,visible:!0,cls:"",area:!1}),this.setElement(this._template(this.options)),void 0!==this.options.value&&this.value(this.options.value),this.options.disabled&&this.$el.prop("disabled",!0),this.options.visible||this.$el.hide();var c=this;this.$el.on("input",function(){c.options.onchange&&c.options.onchange(c.$el.val())})},value:function(a){return void 0!==a&&this.$el.val(a),this.$el.val()},_template:function(a){return a.area?'<textarea id="'+a.id+'" class="ui-textarea '+a.cls+'"></textarea>':'<input id="'+a.id+'" type="'+a.type+'" value="'+a.value+'" placeholder="'+a.placeholder+'" class="ui-input '+a.cls+'">'}}),o=Backbone.View.extend({initialize:function(a){this.options=a,this.setElement(this._template(this.options)),void 0!==this.options.value&&this.value(this.options.value)},value:function(a){return void 0!==a&&this.$("hidden").val(a),this.$("hidden").val()},_template:function(a){var b='<div id="'+a.id+'" >';return a.info&&(b+="<div>"+a.info+"</div>"),b+='<hidden value="'+a.value+'"/></div>'}});return{Anchor:k,Button:f.ButtonDefault,ButtonIcon:f.ButtonIcon,ButtonCheck:f.ButtonCheck,ButtonMenu:f.ButtonMenu,ButtonLink:f.ButtonLink,Icon:j,Image:h,Input:n,Label:i,Message:l,Modal:g,RadioButton:d.RadioButton,Checkbox:d.Checkbox,Radio:d.Radio,Searchbox:m,Select:b,Hidden:o,Slider:c,Drilldown:e}});
 //# sourceMappingURL=../../../maps/mvc/ui/ui-misc.js.map
